Re-arranged and cleaned up the codebase.
