+++ Native/Messages/Abstractions/IBaseMessageReceiver.cs
@@ -1,4 +1,4 @@
-﻿namespace Chopsticks.Messages
+﻿namespace Chopsticks.Messages.Abstractions
 {
     public interface IBaseMessageReceiver<TMessage>
     {
