+++ Native/Messages/Abstractions/IMessageReceiver.cs
@@ -1,4 +1,4 @@
-﻿namespace Chopsticks.Messages
+﻿namespace Chopsticks.Messages.Abstractions
 {
     public interface IMessageReceiver<TMessage, TCollective> :
         IBaseMessageReceiver<TMessage>,
