--- Native/Messages/IRegisteredSignalHandler.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Chopsticks.Messages
-{
-    public interface IRegisteredSignalHandler<TSignal, TAsync> : IDisposable
-    {
-        public TAsync RegisteredHandle();
-
-        // TODO :: Settings.
-    }
-}
