+++ Native/Messages/TaskBased/Commands/IAsyncTaskCommandReceiver.cs
@@ -1,0 +1,10 @@
+﻿using Chopsticks.Messages.Abstractions;
+using System.Threading.Tasks;
+
+namespace Chopsticks.Messages.TaskBased.Commands
+{
+    public interface IAsyncTaskCommandReceiver<TMessage> :
+        IAsyncMessageReceiver<TMessage, Task, CollectiveTaskCommandReceiver<TMessage>>
+    {
+    }
+}
