+++ Native/Messages/TaskBased/Commands/ITaskCommandReceiver.cs
@@ -1,0 +1,9 @@
+﻿using Chopsticks.Messages.Abstractions;
+
+namespace Chopsticks.Messages.TaskBased.Commands
+{
+    public interface ITaskCommandReceiver<TMessage> :
+        IMessageReceiver<TMessage, CollectiveTaskCommandReceiver<TMessage>>
+    {
+    }
+}
